+++ src/aioredis_cluster/typedef.py
@@ -7,7 +7,7 @@ from aioredis_cluster.abc import AbcConnection, AbcPool
 
 TRedis = TypeVar("TRedis", bound=Redis)
 
-
+BytesOrStr = Union[bytes, str]
 AioredisAddress = Union[str, Tuple[str, int]]
 SlotsResponse = List[List]
 CommandsFactory = Callable[[AbcConnection], TRedis]